ai_pipeline/requirements.txt
# Core
sqlalchemy==2.0.28
pyodbc==5.1.0
python-dotenv==1.0.1

# Data & ML
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
xgboost==2.0.3
joblib>=1.3.0

# Azure OCR + LLM Pipeline
azure-ai-formrecognizer>=3.3.0
openai>=1.10.0
pydantic>=2.5.0

# Frontend
streamlit==1.32.2

# Datos sintéticos
Faker==24.4.0
